+++ Client/src/pages/register.aspx.js
@@ -71,6 +71,29 @@ constructor(props){
   };
 }
 
+ onRegisterClick() {
+  window.TwoCoInlineCart.config.app.merchant = 'ONLMETEC';
+  window.TwoCoInlineCart.config.app.iframeLoad = 'immediate';
+  window.TwoCoInlineCart.config.cart.host = 'https:\/\/secure.2checkout.com';
+  window.TwoCoInlineCart.config.cart.customization = 'ONLMETEC-inline-one-step';
+  window.TwoCoInlineCart.setup.setIframeLoad('immediate'); 
+  window.TwoCoInlineCart.register();
+  
+  window.TwoCoInlineCart.billing.setData({
+    name: '', 
+    email: '', 
+    country: '',
+  });
+
+  window.TwoCoInlineCart.products.removeAll(); 
+
+  window.TwoCoInlineCart.products.add({
+    code: "604132"
+  });
+  
+  window.TwoCoInlineCart.cart.checkout();
+}
+
 componentDidMount(){
   const cookies = new Cookies();
 
@@ -89,6 +112,8 @@ componentDidMount(){
  })
 }
 
+
+
 render(){
     return (
       <Layout 
@@ -100,7 +125,8 @@ render(){
        metaKeywords=""
       >
         <Helmet>
-        {<script type='text/javascript'>{`
+
+          {<script type='text/javascript'>{`
                   window._vwo_code = window._vwo_code || (function(){
                   var account_id=279977,
                   settings_tolerance=5000,
@@ -323,23 +349,7 @@ render(){
               margin-top: 15px;
             }
           }`}</style>}
-          {<script>{`
-          (function (document, src, libName, config) {
-            var script             = document.createElement('script');
-            script.src             = src;
-            script.async           = true;
-            var firstScriptElement = document.getElementsByTagName('script')[0];
-            script.onload          = function () {
-            for (var namespace in config) {
-            if (config.hasOwnProperty(namespace)) {
-              window[libName].setup.setConfig(namespace, config[namespace]);
-              }
-            }
-            window[libName].register();
-            };
-            firstScriptElement.parentNode.insertBefore(script, firstScriptElement);
-            })(document, 'https://secure.avangate.com/checkout/client/twoCoInlineCart.js', 'TwoCoInlineCart',{"app":{"merchant":"ONLMETEC","iframeLoad":"immediate"},"cart":{"host":"https:\/\/secure.2checkout.com","customization":"inline"}});
-            `}</script>}
+          {<script type='text/javascript' src='https://secure.2checkout.com/checkout/client/twoCoInlineCart.js' onLoad={this.handleScriptLoad} />}
         </Helmet>
         
         <div className="screen-wrapper first">
@@ -386,7 +396,7 @@ render(){
                 <Text className="limited-offer-text">{this.props.t("Time limited offer")}</Text>
                 <LstDay  MText = {"till " + mounth[currentMounth] + " " + getLastDayOfMonth(currentYear, currentMounth) + ", " +  currentYear} />
                 <Button                
-                href={this.state.hrefUnlim}
+                  onClick={this.onRegisterClick}
                   backgroundColor="orange"
                   color="#ffffff"
                   className="buy-block-button"